Add sampler to build root span context
The current behavior for extracting OpenTelemetry `SpanContext`
instances from tracing `Span`s is to always sample if there is no parent
(which is not the expected behavior). This patch instead stores a copy
of the configured OpenTelemetry `api::Sampler` instance in the layer and
uses that to properly sample requests before setting trace flags.