Start span to build root span context
The current behavior for extracting OpenTelemetry `SpanContext`
instances from tracing `Span`s is to always sample if there is no parent
(which is not the expected behavior). This patch instead starts
OpenTelemetry spans when a span context is requested to properly sample
before setting trace flags.

In order to accomplish this, the `OpenTelemetryLayer` accepts global
boxed tracers so `WithContext` can accept the necessary struct without
requiring a generic parameter.